// dashboard/src/generated/operations.ts
export const operations: Operation[] = [
// ...
  {
// ...
    "responses": {
      "200": {
        "type": "object",
        "required": [
          "players"
        ],
        "properties": {
          "players": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "uuid",
                "name",
                "dimension",
                "x",
                "y",
                "z",
                "inventory"
              ],
              "properties": {
// ...
                "inventory": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "slot": {
                        "type": "integer"
                      },
                      "itemId": {
                        "type": "string"
                      },
                      "count": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
// ...
    }
  },
// ...
] satisfies Operation[];